app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kavisek <kavi.skhon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.0"
black = "^21.7b0"
matplotlib = "^3.4.2"
isort = "^5.9.2"
psycopg2-binary = "^2.9.1"
ipykernel = "^6.0.2"
google-cloud-storage = "^1.41.0"
google-cloud-secret-manager = "^2.6.0"
google-cloud-pubsub = "^2.6.1"
google-cloud-bigquery = {version = "^2.21.0", python = ">=3.8,<3.10"}
python-dotenv = "^0.18.0"
pytest = "^5.2"
SQLAlchemy = "^1.4.21"
flake8 = "^3.9.2"
Jinja2 = "^3.0.1"
PyYAML = "^5.4.1"
pytest-black = "^0.3.12"
pytest-cov = "^2.12.1"
sendgrid = "^6.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
